+++ examples/storybook/stories/Wrapper/RicosBasicUsageStory.js
@@ -11,12 +11,12 @@ import {
 import { pluginDivider } from 'wix-rich-content-plugin-divider/dist/module.viewer';
 
 import dividerContentState from '../../../../e2e/tests/fixtures/divider.json';
-import DividerEditor from './WrapperBasicUsage';
-import sourcecode from '!!raw-loader!./WrapperBasicUsage.js';
+import DividerEditor from './RicosBasicUsage';
+import sourcecode from '!!raw-loader!./RicosBasicUsage.js';
 
 export default () => {
   return (
-    <Page title="Wrapper Basic Usage">
+    <Page title="Ricos Basic Usage">
       <Section type={Section.Types.COMPARISON}>
         <RichContentEditorBox sourcecode={sourcecode} content={dividerContentState}>
           <DividerEditor content={dividerContentState} />