+++ examples/storybook/stories/Wrapper/index.js
@@ -1,11 +1,11 @@
 import { storiesOf } from '@storybook/react';
-import BasicUsageStory from './WrapperBasicUsageStory';
+import BasicUsageStory from './RicosBasicUsageStory';
 import ExampleApp from './ExampleAppStory';
 import ThemesStory from './Themes';
-import StaticToolbarWraper from './WrapperStaticToolbar';
+import RicosStaticToolbar from './RicosStaticToolbar';
 
-storiesOf('Rich Content Wrapper', module)
+storiesOf('Ricos', module)
   .add('Basic Usage', BasicUsageStory)
   .add('Example App', ExampleApp)
   .add('Palettes', ThemesStory)
-  .add('Static Toolbar', StaticToolbarWraper);
+  .add('Static Toolbar', RicosStaticToolbar);